feat: 기본 security, h2-console, swagger 설정 추가
